+++ src/objects/services/storage/notify_message.rs
@@ -1,41 +1,31 @@
-use std::collections::HashMap;
+use crate::primitives::{NotifyType, NotifyTypeImpl, Unit};
+use pyo3::prelude::*;
 use avro_rs::types::Value;
 use log::warn;
-use pyo3::prelude::*;
-use crate::protocol::primitives::{ElementType, Unit};
-use crate::protocol::avro::{Builder, ProtocolMessage, UNIT_ELEMENT_MESSAGE_SCHEMA};
-use crate::protocol::objects::{FromProtocolMessage, ToProtocolMessage};
-use crate::utils::value_to_string;
+use crate::avro::{Builder, NOTIFY_MESSAGE_SCHEMA, ProtocolMessage};
+use crate::objects::{FromProtocolMessage, ToProtocolMessage};
 
 #[derive(Debug, Clone, PartialEq)]
 #[pyclass]
-pub struct UnitElementMessage {
+pub struct NotifyMessage {
     #[pyo3(get, set)]
     pub stream_unit: Unit,
     #[pyo3(get, set)]
-    pub element: ElementType,
+    pub saved_ms: u64,
     #[pyo3(get, set)]
-    pub value: Vec<u8>,
-    #[pyo3(get, set)]
-    pub attributes: HashMap<String, String>,
-    #[pyo3(get, set)]
-    pub last: bool,
+    pub notify_type: NotifyType,
 }
 
 #[pymethods]
-impl UnitElementMessage {
+impl NotifyMessage {
     #[new]
     pub fn new(stream_unit: Unit,
-               element: ElementType,
-               value: Vec<u8>,
-               attributes: HashMap<String, String>,
-               last: bool) -> Self {
-        UnitElementMessage {
+               saved_ms: u64,
+               notify_type: NotifyType) -> Self {
+        NotifyMessage {
             stream_unit,
-            element,
-            value,
-            attributes,
-            last,
+            saved_ms,
+            notify_type,
         }
     }
 
@@ -51,20 +41,18 @@ impl UnitElementMessage {
     const __hash__: Option<Py<PyAny>> = None;
 }
 
-impl FromProtocolMessage for UnitElementMessage {
+impl FromProtocolMessage for NotifyMessage {
     fn load(message: &ProtocolMessage) -> Option<Self> where Self: Sized {
-        if message.schema != UNIT_ELEMENT_MESSAGE_SCHEMA {
+        if message.schema != NOTIFY_MESSAGE_SCHEMA {
             return None;
         }
-
         match &message.object {
             Value::Record(fields) => match fields.as_slice() {
                 [
                 (_, Value::Record(stream_unit_fields)),
-                (_, Value::Long(element)),
-                (_, Value::Bytes(value)),
-                (_, Value::Map(attributes)),
-                (_, Value::Boolean(last))
+                (_, Value::Int(last_element)),
+                (_, Value::Long(saved_ms)),
+                (_, Value::Enum(_index, notify_type)),
                 ] => match stream_unit_fields.as_slice() {
                     [
                     (_, Value::Bytes(stream_name)),
@@ -72,12 +60,14 @@ impl FromProtocolMessage for UnitElementMessage {
                     (_, Value::Enum(_index, track_type)),
                     (_, Value::Long(unit))
                     ] => {
-                        Some(UnitElementMessage {
+                        Some(NotifyMessage {
                             stream_unit: Unit::new(stream_name.clone(), track_name.clone(), track_type.clone(), *unit),
-                            element: element.clone() as i16,
-                            value: value.clone(),
-                            attributes: attributes.iter().map(|x| (x.0.clone(), value_to_string(x.1).or(Some(String::from(""))).unwrap())).collect(),
-                            last: last.clone(),
+                            saved_ms: *saved_ms as u64,
+                            notify_type: match notify_type.as_str() {
+                                "READY" => NotifyType::ready(last_element.clone() as i16),
+                                "NEW" => NotifyType::new(),
+                                _ => NotifyType { obj: NotifyTypeImpl::NotImplemented }
+                            },
                         })
                     }
                     _ => {
@@ -86,7 +76,7 @@ impl FromProtocolMessage for UnitElementMessage {
                     }
                 }
                 _ => {
-                    warn!("Unable to match AVRO Record to to UnitElementMessage");
+                    warn!("Unable to match AVRO Record to to NotifyMessage");
                     None
                 }
             }
@@ -98,17 +88,30 @@ impl FromProtocolMessage for UnitElementMessage {
     }
 }
 
-impl ToProtocolMessage for UnitElementMessage {
+impl ToProtocolMessage for NotifyMessage {
     fn save(&self, mb: &Builder) -> Option<ProtocolMessage> {
-        let mut obj = mb.get_record(UNIT_ELEMENT_MESSAGE_SCHEMA);
-        obj.put("stream_unit", self.stream_unit.to_avro_record());
-        obj.put("element", Value::Long(self.element.into()));
-        obj.put("value", Value::Bytes(self.value.clone()));
-        obj.put("attributes", Value::Map(self.attributes.iter().map(|x| (x.0.clone(), Value::String(x.1.clone()))).collect()));
-        obj.put("last", Value::Boolean(self.last));
+        let mut obj = mb.get_record(NOTIFY_MESSAGE_SCHEMA);
+        obj.put(
+            "stream_unit",
+            self.stream_unit.to_avro_record(),
+        );
+        obj.put("saved_ms", Value::Long(self.saved_ms as i64));
+        match &self.notify_type.obj {
+            NotifyTypeImpl::Ready(elt) => {
+                obj.put("notify_type".into(), Value::Enum(0, "READY".into()));
+                obj.put("last_element".into(), Value::Int(*elt as i32));
+            }
+            NotifyTypeImpl::New => {
+                obj.put("notify_type".into(), Value::Enum(1, "NEW".into()));
+                obj.put("last_element".into(), Value::Int(-1));
+            }
+            NotifyTypeImpl::NotImplemented => {
+                panic!("Unable to handle unsupported NotifyType");
+            }
+        }
 
         Some(ProtocolMessage {
-            schema: String::from(UNIT_ELEMENT_MESSAGE_SCHEMA),
+            schema: String::from(NOTIFY_MESSAGE_SCHEMA),
             object: Value::from(obj),
         })
     }
@@ -116,29 +119,23 @@ impl ToProtocolMessage for UnitElementMessage {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashMap;
     use uuid::Uuid;
     use crate::protocol::avro::Builder;
-    use crate::protocol::objects::{FromProtocolMessage, ToProtocolMessage};
-    use crate::protocol::objects::services::storage::unit_element_message::UnitElementMessage;
-    use crate::protocol::primitives::{pack_stream_name, pack_track_name, Unit};
+    use crate::objects::{FromProtocolMessage, ToProtocolMessage};
+    use crate::objects::services::storage::notify_message::NotifyMessage;
+    use crate::protocol::primitives::{NotifyType, pack_stream_name, pack_track_name, Unit};
     use crate::utils::get_avro_path;
 
-    #[test]
-    fn test_load_save_req() {
+    fn test_load_save_req_int(notify_type: NotifyType) {
         let mb = Builder::new(get_avro_path().as_str());
 
         let track_name = pack_track_name(&String::from("test")).unwrap();
         let stream_uuid = Uuid::parse_str("fa807469-fbb3-4f63-b1a9-f63fbbf90f41").unwrap();
         let stream_name = pack_stream_name(&stream_uuid);
 
-        let req = UnitElementMessage::new(
+        let req = NotifyMessage::new(
             Unit::new(stream_name.to_vec(), track_name.to_vec(), String::from("VIDEO"), 3),
-            2,
-            vec![0, 1],
-            HashMap::from([("a".into(), "b".into()), ("c".into(), "d".into())]),
-            true,
-        );
+            0, notify_type);
 
         let req_envelope_opt = req.save(&mb);
         assert!(req_envelope_opt.is_some());
@@ -151,12 +148,18 @@ mod tests {
 
         let req_envelope = req_envelope_opt.unwrap();
 
-        let new_req_opt = UnitElementMessage::load(&req_envelope);
+        let new_req_opt = NotifyMessage::load(&req_envelope);
 
         assert!(new_req_opt.is_some());
 
         let new_req = new_req_opt.unwrap();
 
         assert_eq!(req, new_req);
+    }
+
+    #[test]
+    fn test_load_save_req() {
+        test_load_save_req_int(NotifyType::new());
+        test_load_save_req_int(NotifyType::ready(100));
     }
 }
