--- src/protocol/mod.rs
@@ -1,6 +1,0 @@
-pub mod primitives;
-pub mod avro;
-pub mod objects;
-
-
-
